+++ core/components/Buttons/CopyToClipboardButton.tsx
