+++ core/components/Hero/Hero.tsx
@@ -1,8 +1,7 @@
-import { styled } from '@maximeheckel/design-system';
+import { H1, styled } from '@maximeheckel/design-system';
 import Image from 'next/image';
 import React from 'react';
 import { loader } from '../../../lib/next-image-loader';
-import { H1 } from '../Typography';
 interface HeroImgProps {
   src: string;
   className: string;
