+++ core/components/MDX/Widgets/DesignSystem/ColorTokens.tsx
@@ -3,11 +3,11 @@ import {
   Button,
   Card,
   Flex,
+  Grid,
   Icon,
   Text,
 } from '@maximeheckel/design-system';
-import LightDarkSwitcher from '@theme/components/Button/LightDarkSwitcher';
-import Grid from '@theme/components/Grid';
+import LightDarkSwitcher from '@theme/components/Buttons/LightDarkSwitcher';
 import Tooltip from '@theme/components/Tooltip';
 import { useTheme } from '@theme/context/ThemeContext';
 import { AnimatePresence, motion } from 'framer-motion';
