+++ core/components/MDX/Widgets/DesignSystem/FlexGridShowcase.tsx
@@ -1,6 +1,5 @@
-import { Box, Card, Flex } from '@maximeheckel/design-system';
+import { Box, Card, Flex, Grid } from '@maximeheckel/design-system';
 import { HighlightedCodeText } from '@theme/components/Code/CodeBlock';
-import Grid from '@theme/components/Grid';
 import React from 'react';
 
 const StyledBox = () => (
