--- core/components/Typography/constant.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_TAG = 'h1';
