--- core/components/Typography/index.tsx
@@ -1,4 +1,0 @@
-export { default, EM, Strong } from './Text';
-export { default as Heading, H1, H2, H3, H4 } from './Heading';
-
-export type { HeadingProps, ShortHandHeadingProps } from './types';
