--- core/components/VisuallyHidden/index.tsx
@@ -1,1 +1,0 @@
-export { VisuallyHidden as default } from './VisuallyHidden';
