+++ pages/404.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable react/no-unescaped-entities */
-import { styled, Anchor, H1, H3 } from '@maximeheckel/design-system';
-import Grid from '@theme/components/Grid';
+import { styled, Anchor, Grid, H1, H3 } from '@maximeheckel/design-system';
 import { DefaultSeo } from '@theme/components/Seo';
 import Link from 'next/link';
 import siteConfig from '../config/site';
