+++ public/rss.xml
@@ -9,7 +9,7 @@
             <link>https://blog.maximeheckel.com</link>
         </image>
         <generator>RSS for Node</generator>
-        <lastBuildDate>Fri, 20 May 2022 22:59:31 GMT</lastBuildDate>
+        <lastBuildDate>Fri, 20 May 2022 23:35:40 GMT</lastBuildDate>
         <atom:link href="https://blog.maximeheckel.com/rss.xml" rel="self" type="application/rss+xml"/>
         <language><![CDATA[en]]></language>
         <item>
